requirements.txt
fastapi==0.104.1
uvicorn==0.23.2

azure-cognitiveservices-speech==1.48.2
azure-core==1.39.0

numpy==1.24.3

soundcard==0.4.3
sounddevice==0.4.6

httpx==0.28.1
requests==2.32.5

pydantic==2.12.5

python-dotenv==1.2.2

mysql-connector-python==9.6.0

tqdm==4.67.3